+++ R/last.R
@@ -5,27 +5,29 @@
 #' object of `MCMC` is an object of class [coda::mcmc], this is a way to capture
 #' more information in case the user needs it.
 #' 
-#' @name last-mcmc
-#' @return `last_*` returns the corresponding variable.
+#' @name fmcmc-info
+#' @return `get_*` returns the corresponding variable passed to the last call
+#' of [MCMC].
 #' @export
-LAST_RUN <- structure(
+MCMC_INFO <- structure(
   list2env(list(
     data.   = list(),
     ptr     = NULL,
     i       = NA_integer_,
     nchains = 0L
   ), envir = new.env()),
-  class = c("fmcmc_run", "environment"))
+  class = c("fmcmc_info", "environment"))
 
-#' @noRd Clears the LAST_RUN environment and sets the number of chains
+#' Clears the MCMC_INFO environment and sets the number of chains
 #' @param nchains an integer
 #' @param env an optional environment (this is useful in the context of parallel chains.)
-LAST_RUN$clear <- function(nchains, env) {
+#' @noRd 
+MCMC_INFO$clear <- function(nchains, env) {
   
-  lapply(LAST_RUN$data., function(d) rm(list = ls(all.names = TRUE, envir = d), envir = d))
+  lapply(MCMC_INFO$data., function(d) rm(list = ls(all.names = TRUE, envir = d), envir = d))
   
   if (missing(env))
-    LAST_RUN$data. <- replicate(
+    MCMC_INFO$data. <- replicate(
       nchains, list2env({
         list(logpost = numeric())
       }
@@ -35,39 +37,41 @@ LAST_RUN$clear <- function(nchains, env) {
     if (nchains != 1L)
       stop("When clearing with an environment, nchains should be 1.")
     
-    LAST_RUN$data. <- list(env)
+    MCMC_INFO$data. <- list(env)
     
   }
     
-  LAST_RUN$nchains <- nchains
-  LAST_RUN$info    <- new.env()
+  MCMC_INFO$nchains <- nchains
+  MCMC_INFO$info    <- new.env()
   invisible()
   
 }
 
-#' @noRd This function sets the current pointer (useful when running in
+#' This function sets the current pointer (useful when running in
 #' serial fashion.)
-LAST_RUN$set_ptr <- function(i) {
+#' @noRd 
+MCMC_INFO$set_ptr <- function(i) {
   
-  if (i > LAST_RUN$nchains)
-    stop("fmcmc_run pointer out of range.", call. = FALSE)
+  if (i > MCMC_INFO$nchains)
+    stop("fmcmc_info pointer out of range.", call. = FALSE)
   
-  LAST_RUN$ptr <- LAST_RUN$data.[[i]]
-  LAST_RUN$i   <- i
+  MCMC_INFO$ptr <- MCMC_INFO$data.[[i]]
+  MCMC_INFO$i   <- i
   
   invisible()
   
 }
 
-#' @noRd Combine
-LAST_RUN$c_ <- function(x, val) {
+#' Combine
+#' @noRd
+MCMC_INFO$c_ <- function(x, val) {
   
-  assign(x = x, value = c(LAST_RUN$ptr[[x]], val), envir = LAST_RUN$ptr)
+  assign(x = x, value = c(MCMC_INFO$ptr[[x]], val), envir = MCMC_INFO$ptr)
   
 }
 
 #' @export
-print.fmcmc_run <- function(x, ...) {
+print.fmcmc_info <- function(x, ...) {
   
   if (x$nchains == 0L)
     cat("-MCMC- has not been called yet. Nothing to show.\n")
@@ -79,21 +83,30 @@ print.fmcmc_run <- function(x, ...) {
   invisible(x)
 }
 
+#' @export
+print.fmcmc_last_mcmc <- function(x, ...) {
+  
+  .Deprecated("MCMC_INFO")
+  print(MCMC_INFO)
+  invisible(x)
+  
+}
+
 MCMC_init <- function(...) {
   
   # Getting the caller environment
-  # LAST_RUN$time_start   <- proc.time()
-  LAST_RUN$time_start <- proc.time()
+  # MCMC_INFO$time_start   <- proc.time()
+  MCMC_INFO$time_start <- proc.time()
   env <- parent.frame()
   
   # Initializing the variables
   for (n in names(env))
     if (n != "...") 
-      assign(n, get(n, envir = env), envir = LAST_RUN$info)
+      assign(n, get(n, envir = env), envir = MCMC_INFO$info)
     
   # Assigning dots
   for (i in seq_len(...length())) {
-    assign(...names()[i], ...elt(i), envir = LAST_RUN$info)
+    assign(...names()[i], ...elt(i), envir = MCMC_INFO$info)
   }
     
   invisible(NULL)
@@ -101,59 +114,32 @@ MCMC_init <- function(...) {
 }
 
 MCMC_finalize <- function() {
-  LAST_RUN$time_end <- proc.time()
+  MCMC_INFO$time_end <- proc.time()
 }
 
 
-
 #' @export
-#' @rdname last-mcmc
-#' @return The function `last_elapsed` returns the elapsed time of the last call
-#' to [MCMC]. In particular, the `MCMC` function records the running time of R
-#' at the beginning and end of the function using [proc.time()]. So this function
-#' returns the difference between the two (`time_end - time_start`).
-last_elapsed <- function() {
-  last_("time_end") - last_("time_start")
-}
-
-#' @export
-#' @rdname last-mcmc
-last_nsteps <- function() last_("nsteps")
-
-#' @export
-#' @rdname last-mcmc
-last_nchains <- function() last_("nchains")
-
-#' @export
-#' @rdname last-mcmc
-last_kernel <- function() last_("kernel")
-
-#' @export
-#' @rdname last-mcmc
-last_conv_checker <- function() last_("conv_checker")
-
-#' @export
-#' @rdname last-mcmc
+#' @rdname fmcmc-info
 #' @param x Character scalar. Name of an argument to retrieve. If `x` was not
 #' passed to the last call, the function returns with an error.
-last_ <- function(x) {
+get_ <- function(x) {
   
-  if (LAST_RUN$nchains == 0L)
+  if (MCMC_INFO$nchains == 0L)
     stop("-MCMC- has not been called yet.", call. = FALSE)
   
   # Checking if it exists outside
-  if (exists(x, envir = LAST_RUN)) {
-    return(LAST_RUN[[x]])
+  if (exists(x, envir = MCMC_INFO, inherits = FALSE)) {
+    return(MCMC_INFO[[x]])
   }
   
-  if (exists(x, envir = LAST_RUN$info)) {
-    return(LAST_RUN$info[[x]])
+  if (exists(x, envir = MCMC_INFO$info, inherits = FALSE)) {
+    return(MCMC_INFO$info[[x]])
   }
   
   # Otherwise, it should be part of the run
-  lapply(LAST_RUN$data., function(env) {
+  lapply(MCMC_INFO$data., function(env) {
     
-    if (!exists(x, envir = env))
+    if (!exists(x, envir = env, inherits = FALSE))
       stop(
         "The object -", x, "- was not found in the last MCMC call.",
         call. = FALSE
@@ -166,12 +152,106 @@ last_ <- function(x) {
 }
 
 #' @export
-#' @rdname last-mcmc
+#' @rdname fmcmc-info
 #' @details The function `get_logpost` returns the `logposterior` value at each
-#' iteration. The values correspond to a named numeric vector.
+#' iteration. The values correspond to a named numeric vector. If `nchains > 1`
+#' then it will return a list of length `nchains` with the corresponding logpost
+#' values for each chain.
+#' @examples 
+#' set.seed(23133)
+#' x <- rnorm(200)
+#' y <- x*2 + rnorm(200)
+#' f <- function(p) {
+#'   sum(dnorm(y - x*p, log = TRUE))
+#' }
+#' 
+#' ans <- MCMC(fun = f, initial = c(0), nsteps=2000)
+#' plot(get_logpost(), type = "l") # Plotting the logpost from the last run
 get_logpost <- function() {
   
-  last_("logpost")
+  if (get_nchains() == 1L)
+    return(get_("logpost")[[1L]])
+  else
+    return(get_("logpost"))
   
 }
 
+
+#' @export
+#' @rdname fmcmc-info
+get_elapsed <- function() {
+  get_("time_end") - get_("time_start")
+}
+
+#' @export
+#' @rdname fmcmc-info
+get_nsteps <- function() get_("nsteps")
+
+#' @export
+#' @rdname fmcmc-info
+get_nchains <- function() get_("nchains")
+
+#' @export
+#' @rdname fmcmc-info
+get_kernel <- function() get_("kernel")
+
+#' @export
+#' @rdname fmcmc-info
+get_conv_checker <- function() get_("conv_checker")
+
+#' @export
+#' @rdname fmcmc-info
+get_seed <- function() get_("seed")
+
+#' @export
+#' @rdname fmcmc-info
+get_burnin <- function() get_("seed")
+
+#' @export
+#' @rdname fmcmc-info
+get_thin <- function() get_("seed")
+
+
+#' @export
+#' @rdname fmcmc-deprecated
+LAST_MCMC <- MCMC_INFO
+class(LAST_MCMC) <- c("fmcmc_last_mcmc", class(LAST_MCMC))
+
+#' Deprecated methods in fmcmc
+#' 
+#' These functions will no longer be included starting version 0.6-0. Instead,
+#' use the function sin [fmcmc-info].
+#' 
+#' @export
+#' @name fmcmc-deprecated
+#' @return The function `last_elapsed` returns the elapsed time of the last call
+#' to [MCMC]. In particular, the `MCMC` function records the running time of R
+#' at the beginning and end of the function using [proc.time()]. So this function
+#' returns the difference between the two (`time_end - time_start`).
+last_elapsed <- function() {
+  last_("time_end") - last_("time_start")
+}
+
+#' @export
+#' @rdname fmcmc-deprecated
+last_nsteps <- function() last_("nsteps")
+
+#' @export
+#' @rdname fmcmc-deprecated
+last_nchains <- function() last_("nchains")
+
+#' @export
+#' @rdname fmcmc-deprecated
+last_kernel <- function() last_("kernel")
+
+#' @export
+#' @rdname fmcmc-deprecated
+last_conv_checker <- function() last_("conv_checker")
+
+#' @export
+#' @rdname fmcmc-deprecated
+#' @param x Name of the object to retrieve.
+last_ <- function(x) {
+  .Deprecated("get_", "The -last_*- methods will be deprecated in the next version of -fmcmc-. Use -get_*- instead.")
+  get_(x)
+}